Fixed review issue: CostAccounting object moved to CostAccounting.scala file
