+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/accounting/HasCost.scala
@@ -1,6 +1,6 @@
 package coop.rchain.rholang.interpreter.accounting
 
-import coop.rchain.rholang.interpreter.CostAccounting.CostStateRef
+import coop.rchain.rholang.interpreter.accounting.CostAccounting.CostStateRef
 
 trait HasCost[F[_]] {
   def cost: CostStateRef[F]
