+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/dispatch.scala
@@ -5,7 +5,7 @@ import cats.effect.{Ref, Sync}
 import coop.rchain.crypto.hash.Blake2b512Random
 import coop.rchain.models.TaggedContinuation.TaggedCont.{Empty, ParBody, ScalaBodyRef}
 import coop.rchain.models._
-import coop.rchain.rholang.interpreter.CostAccounting.CostStateRef
+import coop.rchain.rholang.interpreter.accounting.CostAccounting.CostStateRef
 import coop.rchain.rholang.interpreter.RhoRuntime.RhoTuplespace
 
 trait Dispatch[M[_], A, K] {
